+++ CHANGELOG.md
@@ -17,7 +17,8 @@ x.x.x Release notes (yyyy-MM-dd)
 * Support should realm compact on open callback `Configuration.shouldCompactCallback` as option when configuring a Realm to determine if it should be compacted before being returned.  ([#466](https://github.com/realm/realm-dart/pull/466/))
 * Support ObjectId type. ([#468](https://github.com/realm/realm-dart/pull/468))
 * Support Uuid type. ([#470](https://github.com/realm/realm-dart/pull/470))
-* Support user registration with EmailPassword authentication provider. ([#452](https://github.com/realm/realm-dart/pull/452))
+* Support EmailPassword register user. ([#452](https://github.com/realm/realm-dart/pull/452))
+* Support EmailPassowrd confirm user.
 
 ### Fixed
 * Fixed an issue that would result in the wrong transaction being rolled back if you start a write transaction inside a write transaction. ([#442](https://github.com/realm/realm-dart/issues/442))